Move leftover task event handlers to handlers.py
